Fixed Page Loader to work on every page load
