+++ pages/_app.tsx
@@ -1,11 +1,10 @@
 import 'styles/global.css';
-import { useState } from 'react';
+import { useState, useEffect } from 'react';
 import { ThemeProvider } from 'next-themes';
 import { SessionProvider } from 'next-auth/react';
 
 import { Analytics } from '@vercel/analytics/react';
-
-import Loading from './Loading';
+import Loading from '../components/Loading'
 
 export default function App({
   Component,
@@ -13,12 +12,14 @@ export default function App({
 }) {
   return (
     <SessionProvider session={session}>
-      <ThemeProvider attribute="class">
-        <main className="font-mono">
-          <Component {...pageProps} />
-          <Analytics />
-        </main>
-      </ThemeProvider>
+      <Loading />
+        <ThemeProvider attribute="class">
+          <main className="font-mono">
+            <Component {...pageProps} />
+            <Analytics />
+          </main>
+        </ThemeProvider>
+      
     </SessionProvider>
   );
 }
